+++ src/main/java/com/github/checkit/model/Change.java
@@ -1,5 +1,8 @@
 package com.github.checkit.model;
 
+import com.github.checkit.model.auxilary.AbstractChangeableContext;
+import com.github.checkit.model.auxilary.AbstractEntity;
+import com.github.checkit.model.auxilary.ChangeSubjectType;
 import com.github.checkit.util.TermVocabulary;
 import cz.cvut.kbss.jopa.model.MultilingualString;
 import cz.cvut.kbss.jopa.model.annotations.CascadeType;
