+++ src/main/java/com/github/checkit/model/ProjectContext.java
@@ -1,5 +1,6 @@
 package com.github.checkit.model;
 
+import com.github.checkit.model.auxilary.AbstractEntity;
 import com.github.checkit.util.TermVocabulary;
 import cz.cvut.kbss.jopa.model.annotations.FetchType;
 import cz.cvut.kbss.jopa.model.annotations.OWLClass;
