+++ src/main/java/com/github/checkit/model/Vocabulary.java
@@ -1,5 +1,7 @@
 package com.github.checkit.model;
 
+import com.github.checkit.model.auxilary.AbstractEntity;
+import com.github.checkit.model.auxilary.HasTypes;
 import com.github.checkit.util.TermVocabulary;
 import cz.cvut.kbss.jopa.model.annotations.FetchType;
 import cz.cvut.kbss.jopa.model.annotations.Inferred;
