+++ src/main/java/com/github/checkit/model/auxilary/AbstractChangeableContext.java
@@ -1,4 +1,4 @@
-package com.github.checkit.model;
+package com.github.checkit.model.auxilary;
 
 import com.github.checkit.util.TermVocabulary;
 import cz.cvut.kbss.jopa.model.annotations.MappedSuperclass;
