+++ src/main/java/com/github/checkit/model/auxilary/HasIdentifier.java
@@ -1,4 +1,4 @@
-package com.github.checkit.model;
+package com.github.checkit.model.auxilary;
 
 import java.net.URI;
