+++ src/main/java/com/github/checkit/service/auxiliary/ChangeResolver.java
@@ -3,10 +3,10 @@ package com.github.checkit.service.auxiliary;
 import com.github.checkit.dao.ChangeDao;
 import com.github.checkit.exception.UnexpectedRdfObjectException;
 import com.github.checkit.exception.WrongChangeTypeException;
-import com.github.checkit.model.AbstractChangeableContext;
+import com.github.checkit.model.auxilary.AbstractChangeableContext;
 import com.github.checkit.model.Change;
 import com.github.checkit.model.ChangeObject;
-import com.github.checkit.model.ChangeSubjectType;
+import com.github.checkit.model.auxilary.ChangeSubjectType;
 import com.github.checkit.model.ChangeType;
 import cz.cvut.kbss.jopa.model.MultilingualString;
 import cz.cvut.kbss.jopa.vocabulary.DC;
