+++ src/main/java/com/github/checkit/util/TermVocabulary.java
@@ -31,6 +31,7 @@ public final class TermVocabulary {
     public static final String s_c_odstraneno = CHANGE_DESCRIPTION_NAMESPACE + "odstraněno";
     public static final String s_c_vraceno_zpet = CHANGE_DESCRIPTION_NAMESPACE + "vráceno-zpět";
     public static final String s_c_objekt_zmeny = CHANGE_DESCRIPTION_NAMESPACE + "objekt-změny";
+    public static final String s_c_komentovatelna_entita = CHANGE_DESCRIPTION_NAMESPACE + "komentovatelná-entita";
     public static final String s_p_ma_gestora = CHANGE_DESCRIPTION_NAMESPACE + "má-gestora";
     public static final String s_p_je_gestorem = CHANGE_DESCRIPTION_NAMESPACE + "je-gestorem";
     public static final String s_p_ma_zadatele = CHANGE_DESCRIPTION_NAMESPACE + "má-žadatele";
